pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.4",
    "colorama>=0.4.6",
    "dnspython>=2.7.0",
    "flask>=3.1.1",
    "werkzeug>=3.1.3",
]
